Adding in SD library and some memory optimisations

+++ Button.h
@@ -10,7 +10,7 @@ class Button {
  protected:
    int pin;
    long last_debounce_time;  // the last time the output pin was toggled
-   long debounce_delay;    // the debounce time; increase if the output flickers
+   byte debounce_delay;    // the debounce time; increase if the output flickers
    byte last_button_state;
    
 };
